Add seed Dockerfile to install Carton for bootstrapping
We add a simple Dockerfile that installs Carton for bootstrapping our
Catalyst environment.  We put our Carton-installed modules in `/carton`
rather than in `local`, so as to prevent later issues when users do
`carton install` outside of our Docker container.

We also create a `catalyst` user account in the container for our app to
run as.