Add cpanfile.snapshot
This one is generated from the `docker compose run` command in the
previous commit.  Note that with our setup in the Dockerfile in previous
commits, it is now safe to run `carton install --deployment` from
outside the Docker container, making our app useful even to those who
are not going to run this under Docker.